pyproject.toml
[tool.poetry]
name = "fk-kiltisbot"
version = "0.1.0"
description = "Telegram bot for managing item & account balances, and showing calendar summaries"
authors = ["Fyysikkokilta ry <it@fyysikkokilta.fi>"]
readme = "README.md"
packages = [{include = "kiltisbot"}]

[tool.poetry.dependencies]
python = "^3.13"
google-api-core = '2.25.1'
google-api-python-client = '2.178.0'
google-auth = '2.40.3'
matplotlib = '3.10.5'
pandas = '2.3.1'
python-telegram-bot = {extras = ["job-queue"], version = "^22.3"}
requests = '2.32.4'

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.36.0"
black = "^25.1.0"
pyright = "^1.1.403"
ruff = "^0.12.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.poe.tasks]
_black_check = "black --check ."
_type_check = "pyright ."
_lint = "ruff check ."

lint = ["_black_check", "_lint", "_type_check"]
run.script = "kiltisbot.bot:main()"